+++ packages/jest-types/src/index.ts
@@ -11,5 +11,14 @@ import * as SourceMaps from './SourceMaps';
 import * as TestResult from './TestResult';
 import * as Mocks from './Mocks';
 import * as Transform from './Transform';
+import * as PrettyFormat from './PrettyFormat';
 
-export {Config, Console, SourceMaps, TestResult, Mocks, Transform};
+export {
+  Config,
+  Console,
+  SourceMaps,
+  TestResult,
+  Mocks,
+  Transform,
+  PrettyFormat,
+};
