+++ packages/pretty-format/tsconfig.json
@@ -3,5 +3,9 @@
   "compilerOptions": {
     "rootDir": "src",
     "outDir": "build"
+  },
+  "references": [{
+    "path": "../jest-types"
   }
+  ]
 }
